+++ modules/atac.sh
@@ -5,3 +5,4 @@ MODULE_BIN="atac"
 MODULE_PKG_NAME="atac"
 MODULE_GITHUB="Julien-cpsn/ATAC"
 MODULE_PROJECT_URL="https://github.com/Julien-cpsn/ATAC"
+MODULE_OPTIONAL=true